add AuthApiServices & update di
