+++ app/src/main/java/com/paris_2/san3a/data/repository/AuthRepositoryImpl.kt
@@ -1,8 +1,8 @@
 package com.paris_2.san3a.data.repository
 
 import android.util.Log
+import com.paris_2.san3a.data.service.auth.WhatsAppMessage
 import com.paris_2.san3a.data.source.remote.auth.AuthRemoteDataSource
-import com.paris_2.san3a.data.source.remote.auth.WhatsAppMessage
 import com.paris_2.san3a.data.utils.NetworkConnectionChecker
 import com.paris_2.san3a.domain.NoInternetConnectionException
 import com.paris_2.san3a.domain.RegisterException
