+++ app/src/main/java/com/paris_2/san3a/data/source/remote/auth/AuthRemoteDataSourceImpl.kt
@@ -1,9 +1,12 @@
 package com.paris_2.san3a.data.source.remote.auth
 
-import com.paris_2.san3a.data.source.remote.auth.RetrofitClient.api
+import com.paris_2.san3a.data.service.auth.AuthApiServices
+import com.paris_2.san3a.data.service.auth.WhatsAppMessage
 import com.paris_2.san3a.data.source.remote.auth.dto.OtpDto
 
-class AuthRemoteDataSourceImpl: AuthRemoteDataSource {
+class AuthRemoteDataSourceImpl(
+    private val api: AuthApiServices
+): AuthRemoteDataSource {
     override suspend fun sendMessage(body: WhatsAppMessage): OtpDto {
        return api.sendMessage(body = body)
     }
